requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
duckdb>=1.0.0
python-multipart>=0.0.9
pydantic-settings>=2.0.0
pandas>=2.0.0
pyarrow>=14.0.0

# Build
pyinstaller>=6.8.0

# Dev / Testing
pytest>=8.2.0
httpx>=0.27.0
